reorganize tracker views
